[FIX] sale_subscription: no error when adding deferred aml to invoice
Steps
---
* create a subscription sale order (monthly recurrence, at least one
  product)
* confirm > create invoice
* add a line to the invoice with a start and end date (invisible by
  default)
* confirm invoice, we get a traceback.

Cause
---
This happens because, when `_post`ing account moves, we try to apply the
subscription specific logic to all the move lines, even if there is no
associated sale-order/subscription.
In particular the traceback is thrown when we try to access the
`next_invoice_date` field from a `False` subscription/sale-order.

opw-3940682

closes odoo/enterprise#63512

X-original-commit: 1ad45d52a05afb50d70e68fcff720cf7bc0a4710
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Nathaniel Jacques (naja) <naja@odoo.com>