[FIX] knowledge: fix back button for portal users
When a portal user uses Knowledge, opens articles and presses the back
button, the system updates the url with the previous article id but it
doesn't load the corresponding article.

Steps to reproduce the issue:
1. Log in as a portal user
2. Open an article
3. Click on another article from the sidebar
4. Press the back button
=> The url changes but the system doesn't load the previous article.

This commit resolves this issue by ensuring that the custom web client
created for portal users listens for URL changes and correctly loads
the corresponding article whenever a change is detected. The user should
then be able to navigate within Knowledge using the back button.

task-3743416

X-original-commit: 7008d0d67e1409a7405826fbe1fb8acd8ba16d26
Part-of: odoo/enterprise#63793