[FIX] spreadsheet_edition: use_global_filter visibility
The context menu (and clickable cell) `use_global_filter` should
take the value of the underlying pivot formula, and apply it to the
matching global filters. This works, but was supposed to work only for
`ODOO.PIVOT.HEADER` formulas, and not simple `ODOO.PIVOT` formulas.

This commit fixes the visibility of the `use_global_filter` option
in the context menu, so that it is only visible for `ODOO.PIVOT.HEADER`.

Also removed/changed tests that were testing that the menu was visible
for positional `ODOO.PIVOT` formulas.

closes odoo/enterprise#64226

Task: 3714696
X-original-commit: 746e71c86caf4b458d40759102dfd20a3236c16c
Related: odoo/odoo#168665
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>