[FIX] spreadsheet_edition: use_global_filter visibility
The context menu (and clickable cell) `use_global_filter` should
take the value of the underlying pivot formula, and apply it to the
matching global filters. This works, but was supposed to work only for
`ODOO.PIVOT.HEADER` formulas, and not simple `ODOO.PIVOT` formulas.

This commit fixes the visibility of the `use_global_filter` option
in the context menu, so that it is only visible for `ODOO.PIVOT.HEADER`.

Also removed/changed tests that were testing that the menu was visible
for positional `ODOO.PIVOT` formulas.

closes odoo/enterprise#63316

Task: 3714696
Related: odoo/odoo#166971
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>