[FIX] account_3way_match: don't set other field in compute function
Editing the `amount_currency` field of a product line in the
"Journal Items" tab on an invoice is not really possible at the moment.

On a technical level:
The function `_compute_release_to_pay` assigns fields
`release_to_pay` and `release_to_pay_manual`.
Yet, it is only registered as compute function of `release_to_pay`.
This behaviour is not correct; i.e. it causes a recomputation of the
`amount_currency` field on the line mentioned above.

In this commit a compute function for `release_to_pay_manual`
is introduced. It basically does the job that the old computation
function did.

To reproduce on runbot:
  1. Create an invoice with a single line
  2. Got to "Journal Items"
  3. Edit the "Amount in currency" of the first line (product)
  4. Save the invoice
  5. The "Amount in currency" is reset to the original version.

closes odoo/enterprise#64695

X-original-commit: fba0383def4dc900668f290d657fe5ff00044e71
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>