[FIX] knowledge: fix inconsistent id in article url for portal users
There is an issue with the URL management for Knowledge articles accessed as a
portal user.

Steps to reproduce the issue:
- Open an article as a portal user
- reload the page

Current behavior:
- the accessed url follows the internal pathname scheme : `/odoo/knowledge/{id}`
  instead of the portal pathname scheme (`/knowledge/article/{id}`)
- reloading the page redirects the user to `/my`, the portal home, since the
  servers thinks the user is trying to access an internal view, which is wrong

Expected behavior:
- reloading the page should reload the last visited article, and its id should
  be reflected in the URL, which should follow the knowledge portal pathname
  scheme.

Technical explanation:
When a portal user opens Knowledge, the user loads a page that loads
the backend assets. The client script will then read the article id from
the session and execute an action to open that article.

This commit removes the usage of the session to store the `resId` of visited
articles

Since [1], the `js` router uses a path based routing in the webclient. The
knowledge portal is its own custom webclient and does not follow the same logic
as the standard Odoo webclient.

This commit patches the router with with an altered state management (using
pathnames defined for the portal user, i.e.: `/knowledge/home` and
`/knowledge/article/{id}`, instead of the standard `/odoo/{...}` scheme).

By doing so, we can stop using the URL scheme introduced in [2], which used a
combination of the pathname and the hash:
`/knowledge/article#{id}`, since the hash should be used for link anchors.

[1]: https://github.com/odoo/odoo/commit/c63d14a0485a553b74a8457aee158384e9ae6d3f

[2]: https://github.com/odoo/enterprise/commit/15b9fdd0685a043efa1ba725ed847f2cd20e7558

task-3743416

X-original-commit: 92e8ae7cb5a9d4ed22dbf2b7cafbe92021ff03ab
Part-of: odoo/enterprise#63793