[IMP] account_reports: date selector filter
Date selection is not particularly user-friendly in account reports.

A user may want to quickly navigate between similar periodicity going from one
to the next in a smooth way instead of using calendar.

This PR add a way to navigate between periods using arrows.

Note that we need to remove the `@freeze_time` from the tests because since
we calculate the date filters automatically based on the current date with
javascript if the time is freezed we won't get the correct results.

closes odoo/enterprise#61542

Task-id: 3482119
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>