[REF] account_accountant: simplify code slightly
`attachmentPreviewState.previewEnabled` was more complex than needed:
it was a state solely made to make a getter reactive, due to lack of
observing UI service state reactively with `useState()`.

By putting `useState()` around the ui service in the component,
this let us simplify the code of `previewEnabled` to simplify it
being a getter.

Part-of: odoo/enterprise#63326