[ADD] hr_expense_extract: add sample receipt animation
Add an animation when the user clicks on `Try Sample Receipt`.

We choose this 'original' approach because there was a possibility that we
needed the animation to start (fade in) on a view and finish (fade out) on
another one. At the time of this pr, this looks like the best solution to
allow that. All other solutions that we thought about wouldn't allow that
or some other part of the specifications.

(also we were pressed by time for a freeze and lacked time to look deeply
into other possibilities)

closes odoo/enterprise#63195

Task-id: 3946155
Related: odoo/odoo#167026
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>