[FIX] marketing_automation: test unique utm_source names
With the related community commit whenever utm.source name is written it will
check its uniqueness (generating a unique name if necessary).

[Reproduce]

- Install marketing_automation
- Have Campaign with a Template:
  - Access the Campaign's Template and go to the Settings tab.
  - Note down the NAME of the `utm.source` (found in the Advanced section, under the
    field "Name").
  - Delete this template (which will also delete associated `mailing.mailing` records.)
- Create new template (by returning to the Campaign form, opening an activity,
  creating template)
- Open the Campaign's Template:
  - Disable read-only on the Name field in Advanced section (modify the
    `mailing.mailing.form` view).
  - Change its name to the previously noted `utm.source` NAME.
  - Save -> BUG: Trackback "The name must be unique"

[Why this bug happens?]

When you delete a template (`mailing.mailing`) related to a `utm.source`,
the `utm.source` will still exist. Then, when attempting to change the
`utm.source` name, the system is supposed to detect if the name already
exists and generate a new unique name if necessary. However, due to the
mechanism utilized in the function (ref.1), which searches for existing
names based on existing templates (`mailing.mailing`), which you have deleted,
the existing `utm.source` NAME won't be found. Consequently, the function
mistakenly believes the name is unique and doesn't generate a new name,
leading to the encountered error.

Solution is therefore to check uniqueness on the right model which is
`utm.source` and not `mailing.mailing`.

(ref.1) `UtmMixin._get_unique_names`
https://github.com/odoo/odoo/blob/205caa3d6539187009a8b8bbd1557e75fe045ff3/addons/utm/models/utm_mixin.py#L77

opw-3835704
Task-3837272 (odoo#159699) Incorrect UTM source enumeration

X-original-commit: 2e1cd20c44bb1eb88fb753fab706cf40833101ff
Part-of: odoo/enterprise#64308
Related: odoo/odoo#168799
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Andrzej Pietrusiak <pian@odoo.com>