[FIX] account_accountant: Fix collect_global_info_data function
The aim of this commit is checking that the function
collect_global_info_data properly checks if one of the
user's companies is the one on the journal and not
necessarily the one from the current company (self.env.company).
In the mean time, we're now returning an empty string
instead of a False for the balance_amount's value in the
returned dict because we have a props validation on this
field.

no task id

closes odoo/enterprise#63365

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>