[FIX] l10n_ar_edi: reset to draft button
Before this commit, when sending an invoice to the AFIP and that the result was
accepted, it was still possible to reset the invoice the draft which caused
discrepancies between the data in Odoo and the data that has already been sent
to AFIP. This button to reset to draft is enabled for all users in Accounting.
After an invoice is sent and validated, we cannot edit them.

task-3955111

closes odoo/enterprise#63890

X-original-commit: e28489385a7bcb4a4f075b19deb7d40972a591ff
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>