[FIX] sale_subscription: fix preview template issue in payment reminder
Before this commit, if we try to preview payment reminder template it is giving
error because it is not getting payment link for 'pay now' button.

After this commit, it will not show error on preview because added a fallback
value for 'pay now' button if it not get payment link value.

task-3956795

closes odoo/enterprise#63999

X-original-commit: 6c3c6cb1e646cbbacebb1a06bc132b3eddd8fafd
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>