[FIX] appointment: avoid crash on sparse gantt
When using the new sparse gantt mode [1] users would get a crash
if an appointment was visible.

This is because there was an assumption the rows would always represent
records of the model of the last grouping.

In the sparse mode the rows represent the pill records, in this case events.
So we find the parent group instead to find the relevant attendee, in that mode.

`partner_id` on `calendar.event` in test models is also updated
to use `.id` as `related` doesn't work well with relational fields.

task-3950682

[1] e5f9c54415d3ca10378fc5729b242508396b2316

closes odoo/enterprise#64793

X-original-commit: fe330e9a08bd3ea3296c7e68ddc59ace4c21b28e
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Renaud Thiry (reth) <reth@odoo.com>