[ADD] pos_hr_preparation_display: display employee name on pdis orders
Before on each preprartion display order the user name was displayed
instead of the employee name even when pos_hr module was installed.

This commit adds the employee name to the preparation display order and
when the pos_hr module is installed it will display the employee name
instead of the user name.

closes odoo/enterprise#63422

Taskid: 3956671
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>