[IMP] mrp_mps: remove _compute_bom from wizard
The creation wizard computes the bom from the selected product by
selecting the first bom available.
The issue is that the MPS will add the components of the bom to
the records. If a bom has 100+ components, that's a lot of records.
We want the user to make that a conscious decision instead of an
automated behavior.

closes odoo/enterprise#63423

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>