[FIX] web_grid : Add background to the footer
Steps to reproduce:
	- Install Timesheet module
	- Check the grid view and make sure you have odd number of rows
	- Scroll horizontally and see the row of the column total

Current behavior before PR:
The values of the column total still visible when
scrolling horizontally while it should be hidden.
This is happening because in this commit https://github.com/odoo/enterprise/pull/48406/commits/ec733d78385dee23ca2e338f49c089030b83f4c4 the bg-100 has been
removed so when the row of the column total is not in even position
its cell in the first column won't have a bg in its class
so when values get behind this cell it will still be visible and will
overlay.

Desired behavior after PR is merged:
We added the bg-100 once again so when the row is not
in even position it will still hide the values when
scrolling horizontally

opw-3943960

closes odoo/enterprise#63662

X-original-commit: 01ba25a5f1f94364fb0f4ffcb004403a985e8ed4
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>