[FIX] appointment: fix cache error
Steps to reproduce:

- Install Appointments
- Login as admin (or user with appointment manager rights)
- Go to URL "/appointment"
- Select an appointment and clicks Continue
- Click on previous button and next button

=> ERR_CACHE_MISS (in Chrome) or Document Expired (in Firefox)

The use of "Cache-Control: no-store" leads to an issue when navigating
the browser history through the previous/next buttons.
When posting data from a form, this leads to a security error of the browser
as it prevents the form to be resend automatically.
Instead of submitting the form, we redirect manually when clicking
on the continue button.

Also, to avoid any issue with a previous timezone selected we reused the
timezone from the session before any other if it was set.

task-3945356

closes odoo/enterprise#63458

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>