[FIX] helpdesk: properly filter alias on creation
With the introduction of alias domains in 17.0 an issue was uncovered
in the helpdesk module where any additional alias names with different
domains would be added as followers. This stemmed from an issue in
_ticket_email_split which was not properly filtering emails out that
matched the alias name because self was not yet defined when the method
was called. This was not a problem before because this was immediately
fed into _mail_find_partner_from_emails that would also filter out by
alias_name. This is no longer the case because the domain has to also
match.

Passing in the newly created ticket and using this to map to
team_id.alias_name fixes this issue by properly filtering out alias
emails even with different domains.

opw-3920634

closes odoo/enterprise#63477

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>