[FIX] knowledge: contain the embedded views in their column
This commit fixes an issue with the embedded views inside the columns block in
the editor.
The embedded views were overlapping the columns on their right meaning that some
text could be hidden behind them.

To fix this the margin left and right are now only applied when the embedded
views are directly inside the editor.
If these views are encapsulated inside of another element: columns, tables,
etc., their margins should not be applied.

task-3901864

closes odoo/enterprise#65846

X-original-commit: 841287c31ec57902f541e75ad7b41596ccca3b1f
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>