[IMP] website_generator: send url of the db when fetching the result
In order to inspect the result manually when necessary, we save the url
of the db that received the result of the website scraper.
We do this only when fetching the result because the saas trials cannot
garantee its web.base.url when it is first making the request to the
website scraper. Also updating the tree view for convienience.

Linked to: https://github.com/odoo/iap-apps/pull/808

closes odoo/enterprise#63686

X-original-commit: e7b4a7761df3d17b0cfcdee89915b26528da25a8
Signed-off-by: Romain Derie (rde) <rde@odoo.com>
Signed-off-by: Ivan Klapka (ivkl) <ivkl@odoo.com>