[FIX] l10n_be_hr_contract_salary: show disabled children salary config
Issue:
When configuring an employee before sending the salary configurator,
if we set up children and disabled children for the employee, the option
"disabled children" toggle in the salary configurator does not appear.

Steps To Reproduce:
- on the runbot change the company to "my Belgian company".
- In Laurie Poiret's employee profile, on the "personal information" tab
, change the value for children to 2 and set 1 as disabled.
- Go to Laurie Poiret's contract (stat button) and generate an offer.
- From the offer, navigate to the salary configurator.
- In the salary configurator, proceed until you reach the family
status section.
- The "Disabled children" Toggle does not appear unless you change
the value of the children input.

Solution:
- The `onchangeChildren` method is being invoked in the `start` method
without an `event` argument (`event` is `undefined`).
This results in `event.currentTarget` being `undefined`, causing
the `childCount` variable to be `NaN` and `disabledChildrenNumber` to
get disabled at the end of the method.
- to fix the problem, we added a check for `event` argument in the event
handler, if it is not present, we fallback to getting the children input
value using JQuery and it solves the issue

opw-3925640

closes odoo/enterprise#64501

X-original-commit: 2b72d0329143e11ab6e02933bad8d79968c4608d
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Kawtar Drissi El Bouzaidi (kdeb) <kdeb@odoo.com>