[FIX] timesheet_grid: fix access rights error in timesheet
When a user with 'user' access rigths on the project module that is also
assigned to a task of an internal project, it will prevent the user from
creating a new timesheet from the timesheet grid.

This bug is due to the fact that when a user clicks on the line of the
task, some rpc's call are made on the 'project' model. Since the user
does not have access rigth to the internal project on which the call are
made, an access rigth error is raised.

Step to reproduce:
- create a db with hr_timesheet installed
- have a user with 'user' access rights on project and 'own timesheet'
on timesheet
- create an internal project
- create a task inside that internal project and assigned the user on it
- connect with the other user
- open timesheet app
- select the task from the internal project to start a new timesheet
timer
- an error access right on 'project' is triggered

Solution:
Add a sudo inside the python method 'check can start timer'.

task - 3922511
affected version 16.0 - master

Part-of: odoo/enterprise#68933
Related: odoo/odoo#178342
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>