[FIX] hr_payroll: fix cron interval
Intended:
Generate pdfs is meant to run in every hour.
It was intention as well at creation and it shows from
how nextcall is defined:

<field name="nextcall" eval="(DateTime.now() + timedelta(hours=1))"/>

Current behavior:
But regardless how nextcall is defined, now the cron runs in every month.

This is because interval_type is not explicitly defined on the cron and
thus it takes default value - that is 'months'.

To correct the cron behavior, we explicitly define
 interval_number and interval_type

task-3960037

closes odoo/enterprise#63660

X-original-commit: 279c6cb79235acaadb50f67aea80349af5e0f754
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>