[FIX] account_reports: date filters on end of month
The date filters weren't working on the last day of certain months (31st).

For example, if we are the 31st of may, and we try to get back one month, We
would expect to be the 30th of april. Instead, what the code was doing was
strictly removing one month to the date, so we will be the 31st of april
and since that date doesn't exist, the value that was returned was still
31st of may.

To fix that, we now use luxon to make the date calculation.

closes odoo/enterprise#63658

X-original-commit: 9e1f3a7924d65237abf8db77e871f2db45ea6da6
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>