[IMP] hr_payroll_expense: move creation overhaul 2
This improves the changes made by 6f1f45e1, moving the creation of the
`account.move` at the approval step of an expense report.

Before 6f1f45e1 change:
The `hr_payroll_expense` flow worked by relying on the fact that we
created the expense sheet move at the very last moment, allowing to
set the correct account to its payment term line, to reconcile with
the payslip move easily.

Before this change:
In 6f1f45e1, since approving the report creates draft `account.move`s
& `account.payment`s. This makes it more difficult to reconcile the two
moves. A workaround was found in deleting the draft entry of the
expense sheet when flagging it to be reimbursed through a payslip.

After this change:
This removes that workaround and allows a proper reconciliation of the
two `account.move`s by using the draft entry `matching_number` mechanic.
There are several improvements:
 - Avoid deleting a move, thus allowing it to be edited
 - Automatically reconcile, no matter if the accounts do not match,
   a misc entry is automatically generated if needed
 - Manual fallback, if the amounts of the two `account.move`s
   do not match, we keep the temporary `matching_number` set
   on the lines to help the accountant manually reconcile. We cannot do
   it automatically in this case as it requires user input (selecting
   an account for the write-off entry)

closes odoo/enterprise#67266

Task-id: 3955524
X-original-commit: 59475c4301a9f7f4c29ec1b04455b2a432423ca0
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Julien Alardot (jual) <jual@odoo.com>