[IMP] sale_subscription: anticipate payment on the portal form
Before this commit, when the customer paid on the subscription form in
the portal, the system tried to find an invoice requiring a payment.

It would be an issue if one previous invoice was disputed but the
customer wanted to renew his subscription. The disputed invoice would be
paid and the subscription would not be extended.

subscription_transaction route is still used for invoice payment but it
is now possible to have different behavior depending on the form we
use. After this commit, the customer who pais on the subscription
portal, will automatically create a new invoice.

closes odoo/enterprise#63810

Taskid: XXX
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>