[FIX] account_reports, l10n_mx_reports_closing: period order trial balance
When making comparison in trial balance report,
the values are inverted in relation to the header.

Steps:

- Create and confirm a move X for today and 1000$
- Create and confirm a move Y for today - 1 month and 2000$
- Go to trial balance, select `This month` and one previous period
  comparison in the options
-> Under actual month column, we have the previous month value and vice
   versa

Reason:

In trial Balance we want the period order to be `ascending` by default,
this is why we were inverted the column for this report. However, since
70665c2f40dae9c2823fac520d821e25f4fabb80, the order period is
dynamically set by the `period_order` option in the comparison filter,
and set by default to `ascending`, therefore there is no need to force
the inversion of the columns.

opw-3907491

closes odoo/enterprise#65714

X-original-commit: a839c76a9a9bcef6eb7daf7b76c3a9739a74a037
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>