[FIX] (test_)whatsapp: send selection field value instead of key
Problem:

When a selection field is mapped in a template variable, its key is returned
instead of its value. E.g. a selection field is mapped in a template variable
whose one of the selections is `('out_invoice', 'Customer Invoice')` then
`out_invoice` will be returned instead of `Customer Invoice`.

Technical Problem:

The function `_find_value_from_field_path` of the class `BaseModel` is used to
get the value of the template variable. This function uses '.mapped()` to get
the template's variable value which returns the key instead of the value
whenever the last field of the chain is a selection field.

Solution:

The `last_field_of_chain` and `last_model` are obtained using `field_path`. If
`last_field_of_chain` is of the selection field then the function
`convert_to_export` of the `Selection` class is used to get the value of the
selection field by using the key obtained in the `field_value` and the
`last_model`.

Task-3964078

closes odoo/enterprise#64046

X-original-commit: fda5cf14e97dd46cda3cd9f4b7cc99788e05f8a7
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>