[FIX] l10n_co_dian: neutralize
This commit adds the missing neutralization necessary for the l10n_co_dian
module introduced in [1]

The purpose of the standard neutralization framework is to allow us to
create database copies that will not interact with external systems in
ways that could impact the production database (or if it is not possible
to prevent the interactions, make sure that they are benign or won't
result in actual changes), or impact the customers of the operator of
the production database.

This is mainly useful to allow safe support investigation on database
duplicates.

[1] https://github.com/odoo/enterprise/pull/57925

closes odoo/enterprise#64429

X-original-commit: 8b8f167a66675b6012c8831d07e87a717cc718a7
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>
Signed-off-by: Alexandre Moens (mao) <mao@odoo.com>