[FIX] quality_{control,mrp}: can send to fail loc.
**Current behavior:**
Setting up a control point for a product by quantity with a
failure stock location on manufacturing operations will never
allow a manufactured product that fails the quality check to end
up in the failure location.

**Expected behavior:**
A product created in a manufacturing order can fail a quality
check and end up in the designated failure location in WH.

**Steps to reproduce:**
1. Set up a quality point on some storable product on
     manufacturing operations with:
     `Control per: Quantity`
     `Type: Pass - Fail`
     `Failure Location: Some internal location`

2. Create a new MO with the quality controlled product and
     confirm it

3. Click the Quality Checks button, fail the check (make sure to
     select the faiure location here)

4. After the Quality Check, click Produce All -> Observe that
     the full quantity of products produced by the MO still was
     moved to whatever the default Stock location is, not the
     failure location specified.

**Cause of the issue:**
The method `_move_line_to_failure_location()` expects a picking
and SML to exist, however when the quality check starts from an
MO, we have neither (contrasted with a check starting from a
picking). So the method just skips the logic where the re-route
would occur.

**Fix:**
Override the method in the mrp_quality bridge module, if the
quality check has a production_id then create the SML on the
spot via `MrpProduction.action_confirm()`, then go through the
same logic to split the created line as necessary based on the
failed quantity vs. total quantity of the MO.

opw-3865301

closes odoo/enterprise#62731

Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>