[FIX] spreadsheet_edition: snapshot more frequently
Purpose
-------

Let's say a user works a lot on a spreadsheet and performs a lot of heavy
operations generating very large revisions (with lots of commands), such as
copy-pasting a very large zone (there's one UPDATE_CELL command per copy-pas
cell)

If the user does multiple such very large revision, one after the other,
every thing is fine client side.

The next time the spreadsheet is open though: the server needs to load all
those revisions to send them to the client. The revisions can be so large
it can blow up the server memory limit.
The spreadsheet cannot be open anymore.

Solution
--------

In this commit, we increase the number of snapshot by snapshotting more often.

Previously, we used to snapshot when opening a spreadsheet and no activity
was done in the spreadsheet in the last 12 hours. It's reduced to 2 hours
in this commit.
The reason we don't want to snapshot too often is that it kills the local
history (CTRL+Z) of connected users.

closes odoo/enterprise#63741

Task: 3940465
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>