[FIX] helpdesk: fix activity default filter
Clicking on an Activity group used to set a default filter, in effect of which
user will only see helpdesk tickets with HIS activites that are overdue/today.

This behaviour got temporarily removed and than reintroduced with (ref.1)
(ref.1) achieves that by adding default filters to the search views, it issue
persisted in the helpdesk view becase filters wern't present.

This commits adds filters to the helpdesk's search view so they can be used
by mechanisms (re)introdcuted in (ref.1)

(ref.1)
[FIX] mail: Fix activity default filter issue
https://github.com/odoo/odoo/commit/3d18b9faa1fdcf00d48091ad4be244ca00d33524

[Reproduce steps]
- Install helpdesk
- (optional) create a helpdesk ticket where:
    - current user has a future activity
    - other user has an overdue activity
- open ActivityMenu dropdown (clock icon ".fa-clock-o")
- BUG (ux issue): filters aren't present
    - (if optional): created ticket is shown (beside current user not having any overdue/today activity in there)

opw-3859535

closes odoo/enterprise#63748

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>