[FIX] l10n_mx_edi_sale: remove default on usage
Problem: When the user creates a new quotation from a CRM opportunity,
the quotation does not compute to the partner's usage value and
always default to "General Usage". The default prevents the compute
from getting called.

Purpose: The compute method "_compute_l10n_mx_edi_usage" should be
called when a new quotation is created from the crm opportunity.

Steps to Reproduce on Runbot17:
1. Install Sales, l10n_mx, l10n_mx_edi, CRM, Contacts
2. Set a value for field "Usage" (l10n_mx_edi_usage) on
a partner whose country is Mexico in Contacts
3. Create a CRM opportunity with the partner above
4. Create a new quotation from the opportunity and
observe that the "Usage" value is not the same as
defined on the partner

opw-3934009

closes odoo/enterprise#64528

X-original-commit: c9faa260ea3929f054b56fa7ddd47be36718aea6
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Mylyna Hy (myhy) <myhy@odoo.com>