[FIX] documents_project: project admin should be able to rename project
To reproduce:
=============
- install Project and Documents
- create a project
- give the user the project admin rights and **User** rights for the
  Documents module
- try to rename a project -> access error

Problem:
========
when Document is installed, every project has a related folder in
the Documents. When renaming a project, the folder name should be
updated as well. As the user has no rights to write on Documents,
the renaming fails.

Solution:
=========
set the folder name as `sudo` to avoid access errors

opw-3928018

closes odoo/enterprise#64396

X-original-commit: 82dd838d9ec9c049c41c768401e53b1440cc656d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Abdelouahab Laaroussi (abla) <abla@odoo.com>