[FIX] appointment: prevent attributeerror while clicking on add a leave button
When timezone of user is empty and the user clicks on Add a Leave button,
a traceback will appear.

Steps to reproduce the error:
- Open Profile > Preferences > Select empty in timezone > Save
- Install 'Appointments'
- Go to Appointments > Schedule > Resource Bookings > Add a Leave

Error: A traceback appears:
"AttributeError: 'bool' object has no attribute 'upper'"

https://github.com/odoo/enterprise/blob/7640860f415028d2d516a2df5e4bc26d9d6dd3b4/appointment/wizard/appointment_manage_leaves.py#L14
When the user selects an empty timezone, tz will be False.
So, It will lead to the above traceback.

sentry-5443573507

closes odoo/enterprise#64907

X-original-commit: 68fb3123937e106396132fd534f804ba63177829
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Renilkumar Kajavadra (reka) <reka@odoo.com>