[FIX] l10n_cl_edi: Clear invalid token when verifying claim
Suppose the SII workflow cron job runs, and there are no records
with the `ask_for_status` or `accepted` l10n_cl_dte_status, but there
are a lot with the `claimed` status. If the token on the certificate
happens to be invalid, `_l10n_cl_ask_claim_status` will raise an
InvalidToken exception and abort the cron job. This prevents the records
from being processed. This fix clears an invalid token if it's detected,
so that when the next record is processed a new one will be created. The
record raising the exception should succeed the next time the cron job
runs.

opw-3920273

closes odoo/enterprise#64879

X-original-commit: e11466b9b674ca24168a3b7ca16f62da0b10c994
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Merel Geens <mege@odoo.com>