[FIX] account_asset: reversed entry should be in depreciation schedule
Create an asset
Reverse an entry in the past
=> In depreciation schedule, if we take the current period,
initial amount in Depreciation does not correspond to the cumulative
depreciation in the Depreciation Board of the asset.

The reversed entry are not taken into account, which was normal
before, but now we take reversal entry into account so these should
be too.

The removal of the join revealed that the order of the query was not
enough. The test test_depreciation_schedule_prefix_groups fails
(several assets have the same acquisition_date and accounts).
account_id was added.
opw-3890144

closes odoo/enterprise#64713

X-original-commit: 8d8082b7370906bba2b2c6980f12f30306e4d0b1
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>