[FIX] web_grid: fix timesheet text cut in grid view
Steps to reproduce:
- Install the project and timesheet app.
- Open the project app, create a project, and add a task to it.
- Add a timesheet a week before the current date for the created task.
- Now open the timesheet app in grid view.
- The created project and task are in italic font.
- The last letter of the project and task name is slightly cut.

Issue:
The last letter of the string is slightly cut in the timesheet grid view.

Cause:
The dedicated space allowed for normal text is not sufficient for the italic.
So, the last letter was slightly cut.

Solution:
Provide an extra space to the italic string.

This commit resolves the issue of text cut in the timesheet when the lines are
displayed in italic style.

task-3888277

closes odoo/enterprise#63786

Signed-off-by: Vincent Larcin (vila) <vila@odoo.com>