[IMP] l10n_be_soda: auto update Soda mapping when entry line with Suspense Account is updated
If we upload a Soda and an account is not mapped, we'll fall back on the Suspense Account.

However, when modifying the entry line with the Suspense Account, the general Soda Mapping is not updated.

For UX reasons, we should update the general Soda Mapping too so that the user doesn't have to modify it twice.

However, we shouldn't do the reverse. I.e. when modifying the general Soda Mapping,
we shouldn't modify the existing entries.

closes odoo/enterprise#63812

Taks-id: 3954585
X-original-commit: 52743720736d5466db1300bd71de9a70c749816c
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>