[FIX] l10n_be_reports: _get_bce_number handle legacy values
The aim of this commit is to make the code more robust and defensive by
handling a bad value for the `vat` field.

Context:
Partners having a `partner.vat=False` were making the compact function
crash.

Before the commit:
The flow crash even if the value shouldn't be processed.

After the commit:
The flow is smooth.

closes odoo/enterprise#63979

Task-id: None
X-original-commit: d20b8fd001ab0bebc7036b0a928091d7e9155587
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>