[FIX] mrp_workorder: update qty in MO's of tracked product
Steps to reproduce:

- Create a storable product P tracked by SN
- Create a consumable (or a storable with 5 units on hand) product
  COMP 1 and a storable product COMP 2 (without units on hand)
- Create a BOM for P with an operation op 1 and two component lines:
    - 1 x COMP 1 consumed in op 1
    - 1 x COMP 2 consumed in op 1
- Create and confirm an MO for 5 units of P
- Go to the shopfloor and click on register prodution.
> the qty is updated to 1 on COMP 2 but to 5/1 on COMP 1

As such, if you click on the 5/1, 5 units of COMP 1 will be consumed to
produce only one unit of P

Cause of the issue:

When you confirm the MO, since Comp 1 is a consumable its quantity is
automatically set to 5.0 because reservation are bypassed.
On the other hand, since Comp 2 is a storable without on hand qty,
its quantity stays at 0.0. When you click on register production, or on
the plus sign will trigger a call of the "_set_qty_producing" method.
This call will update the qty_producing of the final product:
https://github.com/odoo/odoo/blob/f86c68ec8340a59407ea9c51dd0ba942f9b4429c/addons/mrp/models/mrp_production.py#L1214-L1218
However, the update of the qty consumed by the raw move will be bypassed
because of these lines:
https://github.com/odoo/odoo/blob/f86c68ec8340a59407ea9c51dd0ba942f9b4429c/addons/mrp/models/mrp_production.py#L1225-L1226
https://github.com/odoo/enterprise/blob/0646022d7726a0cc183b191ca5be4e4bb4368f93/mrp_workorder/models/stock_move.py#L10-L13
And the quantity will therefore not be updated by these lines:
https://github.com/odoo/odoo/blob/f86c68ec8340a59407ea9c51dd0ba942f9b4429c/addons/mrp/models/mrp_production.py#L1228-L1231
However, as the quantity is not set to 0, it will be displayed as
"quantity/should_consume_qty" and clicking on the raw move line will not
update the quantity so 5 units will be marked as consumed ("picked").

opw-3887580

closes odoo/enterprise#68418

X-original-commit: 67a5f40020fca04970145cf84d84adfb7266490d
Related: odoo/odoo#176713
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>
Signed-off-by: William Henrotin (whe) <whe@odoo.com>