[FIX] l10n_be_us_consolidation_demo: make the invoices demo data not raise an error
Steps to reproduce:
- Create a new DB with nothing installed and demo data not active
- Install l10n_be_us_consolidation_demo
- 'Activate the developer mode' in settings page
- 'Load Demo Data' in settings page and confirm
- Check logs and see traceback

Cause:
The account.move is created with the default company and the account.move.line if forced to be created with an account from the company 'l10n_be.demo_company_be'. In the end the move and its lines does not have the same company so the record creation fails in check_company().

Solution:
Specify the company of the account.move in the record to be the same as the lines.

opw-3952634

closes odoo/enterprise#63903

Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>