[FIX] web_enterprise: box-shadow var none instead of 0
In commit odoo/enterprise@aab87024b9dc901dfbca9a9de9739a698989d3f4 we
introduced variables to enable the box-shadows on some bootstrap
components.

However the property was set to `0` which is an invalid property value.
The correct value is `none`[1].

This issue doesn't have any visible impact, so it's safe to be fixed in
master.

task-3755095

[1]: https://developer.mozilla.org/en-US/docs/Web/CSS/box-shadow

closes odoo/enterprise#63904

Signed-off-by: Pierre Paridans (app) <app@odoo.com>