[FIX] account_accountant: Fallback value for digits in bank rec widget
The aim of this commit is giving a fallback value when we need
digits' currency and that we don't have this currency loaded
in the js session.
This commit is almost backport of what we have done in this commit [1].
The only difference is the default value for the toFixed set to 2.

task-3959277

[1]: https://github.com/odoo/enterprise/commit/6a3623c168f3bdf32b9530071e21c014dd19ea07

closes odoo/enterprise#64895

X-original-commit: 361131995e63b93bb1a095e85251817cec091358
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>