[FIX] pos_order_tracking_display: fix translation
Problem
---
On the order status page, the terms translations are ignored.

This happens because it is a 'website=True' routed page, and modules
which create website pages with translations need to let the http
routing know about them by overloading `_get_translations_modules_name`

Note: the page will typically be translated into the browser's preferred
language and not the odoo user's.

opw-3949394

closes odoo/enterprise#64658

X-original-commit: 1095890d3e674f57a17ed8dbf1d6d10bb22f647d
Signed-off-by: Nathaniel Jacques (naja) <naja@odoo.com>