[IMP] l10n_us_payment_nacha: support setting company discretionary data
The Company Discretionary Data field is typically optional, by default
Odoo sets the batch name for traceability. But it could be used by the
bank for their own reasons. From the spec [1]:

> Originator/ODFI may include codes of significance only to them to
> enable specialized handling of all entries within the batch.

Chase instructs their customers to use the field as follows:

> Add your Chase account number in company discretionary data (usually
> on the account set-up screen or ACH/Nacha file setup screen) This is
> the pay-from account for payments or the receiving account for
> collections:
>
> Company discretionary data
> Your Chase account number (Exactly 20 numbers using leading zeroes
> to fill in the gap – e.g. 00000000000123456789)

To support it we add a field on the journal where users can specify
this. Note that we don't zero-pad the number like Chase requests, the
official specification and the BOfA specification [2] indicates this
is an alphanumeric field so we pad with spaces to hopefully be more
compatible. It will be up to the user to manually pad it with zeros if
needed.

This uses an f-string instead of format() because runbot linting now
disallows format().

Based on feedback in opw-3942095.

[1] https://achdevguide.nacha.org/ach-file-details
[2] https://files.nc.gov/ncosc/documents/eCommerce/bank_of_america_nacha_file_specs.pdf

closes odoo/enterprise#63951

Signed-off-by: William André (wan) <wan@odoo.com>