[IMP] l10n_mx_edi: Non-deterministic elements in views
There are currently some elements in views that are not reachable on a
deterministic way when those views are inherited. That means, to reach
them, positional selectors need to be used (e.g. `last()`). That causes
inherited views to get broken when there are small changes in element
position.

A good example of this is [1]:

    <div class="digital-stamp">
        <span>Extra Info</span>
    </div>

Which can't be filtered by:
- Class, because there are several `<div>` elements with the same class
- Text content, because when content is translated, selector will fail

To solve the above, this change provides names for several elements, to make
possible/easier to reach them on a safely manner.

[1] https://github.com/odoo/enterprise/blob/d6346ef8/l10n_mx_edi/views/l10n_mx_edi_report_invoice.xml#L123

closes odoo/enterprise#65180

X-original-commit: 5a7a0567007e0e20fdc312e76ed3c9eaab9bb010
Signed-off-by: Laurent Smet (las) <las@odoo.com>